contacts validations moved to contacts folder
