+++ index.html
@@ -60,13 +60,7 @@
                     </li>
                 </ul>
                 <!--social media-->
-                <div class="social-icon">
-                    <!-- <a href="#" target="_blank" class="fa fa-facebook" aria-hidden="true"></a>
-                    <a href="#" target="_blank" class="fa fa-twitter" aria-hidden="true"></a>
-                    <a href="#" target="_blank" class="fa fa-behance" aria-hidden="true"></a>
-                    <a href="#" target="_blank" class="fa fa-linkedin" aria-hidden="true"></a>
-                    <a href="#" target="_blank" class="fa fa-dribbble" aria-hidden="true"></a> -->
-                </div>
+                <div class="social-icon"></div>
             </div>
         </div>
     </header>
