+++ js/components/contacts/isValidContactName.js
