+++ js/components/contacts/isValidContactText.js
